pyproject.toml
[project]
name = "goit-pythonweb-hw-012"
version = "0.1.0"
description = ""
authors = [
    { name = "MaxymBeregovoi", email = "108676512+maksberegovoi@users.noreply.github.com" }
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "alembic (>=1.17.2,<2.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "fastapi (>=0.124.2,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "fastapi-mail (>=1.6.1,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "cloudinary (>=1.44.1,<2.0.0)",

    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "redis (>=7.1.0,<8.0.0)",
    "redis-lru (>=0.1.2,<0.2.0)"

]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "sphinx (>=9.0.4,<10.0.0)"
]
